fix: flatten checklist components folder. rename checklist items so adhere to file nesting
