+++ src/page/CheckList/components/AlertSensitivityBadge.tsx
