--- src/page/CheckList/components/BulkActions/index.ts
@@ -1,1 +1,0 @@
-export { BulkActionsModal } from './BulkActionsModal';
