--- src/page/CheckList/components/CheckListItem/index.ts
@@ -1,1 +1,0 @@
-export { CheckListItem } from './CheckListItem';
