+++ src/page/CheckList/components/CheckStatusPill.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { useTheme2 } from '@grafana/ui';
 
-import { Pill } from 'page/CheckList/components/CheckListItem/Pill';
+import { Pill } from 'page/CheckList/components/Pill';
 
 interface CheckStatusPillProps {
   enabled: boolean;
