+++ src/page/CheckList/components/ProbesByRegion.tsx
